Don't pluralize package names when possible
