+++ commands/validate.go
@@ -22,7 +22,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
-	"sigs.k8s.io/zeitgeist/dependencies"
+	"sigs.k8s.io/zeitgeist/dependency"
 )
 
 func addValidate(topLevel *cobra.Command) {
@@ -47,7 +47,7 @@ func addValidate(topLevel *cobra.Command) {
 // runValidate is the function invoked by 'addValidate', responsible for
 // validating dependencies in a specified configuration file.
 func runValidate(opts *options) error {
-	client := dependencies.NewClient()
+	client := dependency.NewClient()
 
 	if opts.localOnly {
 		if err := client.LocalCheck(opts.configFile, opts.basePath); err != nil {
